ai/requirements.txt
fastapi
uvicorn
anthropic
psycopg2-binary
redis
python-dotenv
pydantic
